code refinement, from hf fix
