+++ paddlenlp/transformers/tokenizer_utils_base.py
@@ -40,9 +40,9 @@ from huggingface_hub import (
 from huggingface_hub.utils import EntryNotFoundError
 from paddle import __version__
 
+from ..transformers.aistudio_utils import aistudio_download
 from ..utils.downloader import (
     COMMUNITY_MODEL_PREFIX,
-    aistudio_file_exists_download,
     get_path_from_url_with_filelock,
     url_file_exists,
 )
@@ -126,6 +126,20 @@ EncodedInputPair = Tuple[List[int], List[int]]
 SPECIAL_TOKENS_MAP_FILE = "special_tokens_map.json"
 ADDED_TOKENS_FILE = "added_tokens.json"
 TOKENIZER_CONFIG_FILE = "tokenizer_config.json"
+
+
+def aistudio_file_exists_download(repo_id: str, file_path: str):
+    try:
+        cache_path = aistudio_download(
+            repo_id=repo_id,
+            filename=file_path,
+        )
+        return True, cache_path
+    except Exception as e:
+        if "object does not exist" in e.args[0]:
+            return False, None
+        else:
+            raise e
 
 
 def to_py_obj(obj):
@@ -1464,11 +1478,14 @@ class PretrainedTokenizerBase(SpecialTokensMixin):
 
         vocab_files_target = {**cls.resource_files_names, **additional_files_names}
 
-        # From HF Hub or AI Studio
-        if from_hf_hub or from_aistudio:
+        # From HF Hub
+        if from_hf_hub:
+            vocab_files = copy.deepcopy(cls.resource_files_names)
+            vocab_files["tokenizer_config_file"] = cls.tokenizer_config_file
+        # From AI Studio
+        elif from_aistudio:
             # Only include the necessary resource files specified by the tokenizer cls
             # Deep copy to avoid modifiying the class attributes
-            # vocab_files = copy.deepcopy(cls.resource_files_names)
             vocab_files = copy.deepcopy(vocab_files_target)
             vocab_files["tokenizer_config_file"] = cls.tokenizer_config_file
         # From built-in pretrained models
